Bug 1446990 - Safely transmit nsTransparencyMode across IPDL for WinCompositorWidget. r=jrmuizel

+++ widget/windows/CompositorWidgetChild.cpp
@@ -46,7 +46,7 @@ CompositorWidgetChild::OnDestroyWindow()
 void
 CompositorWidgetChild::UpdateTransparency(nsTransparencyMode aMode)
 {
-  Unused << SendUpdateTransparency(static_cast<int32_t>(aMode));
+  Unused << SendUpdateTransparency(aMode);
 }
 
 void
